# docs/swagger.yaml
basePath: /api
definitions:
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  quiz.Quiz:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/user.User'
      created_by_id:
        type: integer
      id:
        type: integer
      question:
        type: string
      quiz_type:
        $ref: '#/definitions/quiz.QuizType'
      selections:
        items:
          $ref: '#/definitions/quiz.QuizSelection'
        type: array
      updated_at:
        type: string
    type: object
  quiz.QuizSelection:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_correct:
        type: boolean
      quiz:
        $ref: '#/definitions/quiz.Quiz'
      quiz_id:
        type: integer
      selection_text:
        type: string
      updated_at:
        type: string
    type: object
  quiz.QuizType:
    enum:
    - single_choice
    - multi_choice
    - true_false
    type: string
    x-enum-varnames:
    - QuizTypeSingleChoice
    - QuizTypeMultiChoice
    - QuizTypeTrueFalse
  quiz_suite.QuizSuite:
    type: object
  user.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing quizzes, quiz suites, and users
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Quizlet API
  version: "1.0"
paths:
  /quiz-suites:
    get:
      description: Retrieve all quiz suites for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/quiz_suite.QuizSuite'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all quiz suites
      tags:
      - quiz-suites
    post:
      consumes:
      - application/json
      description: Create a new quiz suite with the provided details
      parameters:
      - description: Quiz Suite object
        in: body
        name: quiz_suite
        required: true
        schema:
          $ref: '#/definitions/quiz_suite.QuizSuite'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/quiz_suite.QuizSuite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new quiz suite
      tags:
      - quiz-suites
  /quiz-suites/{id}:
    delete:
      description: Delete a quiz suite by its ID
      parameters:
      - description: Quiz Suite ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a quiz suite
      tags:
      - quiz-suites
    get:
      description: Retrieve a specific quiz suite by its ID
      parameters:
      - description: Quiz Suite ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quiz_suite.QuizSuite'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a quiz suite by ID
      tags:
      - quiz-suites
    put:
      consumes:
      - application/json
      description: Update an existing quiz suite with the provided details
      parameters:
      - description: Quiz Suite ID
        in: path
        name: id
        required: true
        type: integer
      - description: Quiz Suite object
        in: body
        name: quiz_suite
        required: true
        schema:
          $ref: '#/definitions/quiz_suite.QuizSuite'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quiz_suite.QuizSuite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a quiz suite
      tags:
      - quiz-suites
  /quizzes:
    get:
      description: Get all quizzes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/quiz.Quiz'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all quizzes
      tags:
      - quizzes
    post:
      consumes:
      - application/json
      description: Create a new quiz with the provided information
      parameters:
      - description: Quiz information
        in: body
        name: quiz
        required: true
        schema:
          $ref: '#/definitions/quiz.Quiz'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/quiz.Quiz'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new quiz
      tags:
      - quizzes
  /quizzes/{id}:
    delete:
      description: Delete a quiz by quiz ID
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a quiz
      tags:
      - quizzes
    get:
      description: Get quiz information by quiz ID
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quiz.Quiz'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a quiz by ID
      tags:
      - quizzes
    put:
      consumes:
      - application/json
      description: Update quiz information by quiz ID
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: integer
      - description: Quiz information
        in: body
        name: quiz
        required: true
        schema:
          $ref: '#/definitions/quiz.Quiz'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quiz.Quiz'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a quiz
      tags:
      - quizzes
  /quizzes/{id}/selections:
    post:
      consumes:
      - application/json
      description: Add a new selection to a quiz
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: integer
      - description: Selection object
        in: body
        name: selection
        required: true
        schema:
          $ref: '#/definitions/quiz.QuizSelection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quiz.Quiz'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a selection to a quiz
      tags:
      - quizzes
  /quizzes/{id}/selections/{selectionId}:
    delete:
      description: Remove a selection from an existing quiz
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: integer
      - description: Selection ID
        in: path
        name: selectionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove a selection from a quiz
      tags:
      - quizzes
  /quizzes/user:
    get:
      description: Get all quizzes created by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/quiz.Quiz'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user's quizzes
      tags:
      - quizzes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"